api/requirements.txt
flask>=3.1.0
flask-cors>=5.0.0
Pillow>=10.4.0
torch>=2.5.0
trimesh>=4.5.0
fast-simplification>=0.1.7
numpy>=1.26.0
# Hunyuan3D-2 (install separately):
#   pip install git+https://github.com/Tencent/Hunyuan3D-2.git
